requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
sentence_transformers
faiss-cpu
mysql-connector-python
